+++ lib/quiver_core.dart
@@ -1,4 +1,6 @@
+export 'src/widgets/date.dart';
 export 'src/widgets/dropdown.dart';
 export 'src/widgets/dropdown_autocomplete.dart';
+export 'src/widgets/month.dart';
 export 'src/widgets/term.dart';
 export 'src/widgets/multiselect_autocomplete.dart';